+++ package.json
@@ -4,7 +4,7 @@
   "version": "1.0.2",
   "private": true,
   "dependencies": {
-    "@brainstack/react": "^1.3.35",
+    "@brainstack/react": "^1.3.46",
     "@fullcalendar/core": "^6.1.4",
     "@fullcalendar/daygrid": "^6.1.4",
     "@fullcalendar/react": "^6.1.4",
